# ORIGINAL/MedicalCenter_PRODUCTION/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Al-Attar Military Medical Center API
  version: 1.0.0
  description: REST API for clinics, exams, and users.
servers:
- url: /api
paths:
...
  /clinics/{id}:
    get:
      summary: Get clinic
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
    put:
      summary: Update clinic
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Updated
    delete:
      summary: Delete clinic
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        '204':
          description: Deleted
...
  /visits/clinic/{clinicId}:
    get:
      summary: List clinic visits (no PII)
      parameters:
      - in: path
        name: clinicId
        schema:
          type: string
        required: true
      responses:
        '200':
          description: OK
...